move scenario state to scenario
