+++ src/interface/src/app/plan/create-scenarios/create-scenarios.component.ts
@@ -30,7 +30,7 @@ import { SetPrioritiesComponent } from './set-priorities/set-priorities.componen
 import { ConstraintsPanelComponent } from './constraints-panel/constraints-panel.component';
 import { GoalOverlayService } from './goal-overlay/goal-overlay.service';
 import { canAddTreatmentPlan } from '../permissions';
-import { ScenarioState } from 'src/app/maplibre-map/scenario.state';
+import { ScenarioState } from 'src/app/scenario/scenario.state';
 import { MatTabGroup } from '@angular/material/tabs';
 import { DataLayersStateService } from '../../data-layers/data-layers.state.service';
 import { PlanState } from '../plan.state';
