+++ src/interface/src/app/plan/create-scenarios/set-priorities/set-priorities.component.spec.ts
@@ -14,7 +14,7 @@ import { TreatmentGoalsService } from '@services';
 import { SetPrioritiesComponent } from './set-priorities.component';
 import { Component } from '@angular/core';
 import { MockProvider } from 'ng-mocks';
-import { ScenarioState } from '../../../maplibre-map/scenario.state';
+import { ScenarioState } from '../../../scenario/scenario.state';
 import { SectionComponent } from '../../../../styleguide/collapsible-panel/section.component';
 
 @Component({ selector: 'app-scenario-tooltip', template: '' })
