+++ src/interface/src/app/plan/plan.component.ts
@@ -20,7 +20,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { BreadcrumbService } from '@services/breadcrumb.service';
-import { ScenarioState } from '../maplibre-map/scenario.state';
+import { ScenarioState } from '../scenario/scenario.state';
 import { getPlanPath } from './plan-helpers';
 import { PlanState } from './plan.state';
 import { canAddScenario } from './permissions';
