+++ src/interface/src/app/scenario/scenario.state.spec.ts
