+++ src/interface/src/app/scenario/scenario.state.ts
